(1) add a test for multi variant and header VCF file parsing; (2) some refactoring.
