+++ GeUtilities.Tests/Model/ParsedIntervals.cs
